+++ cocos2d/TextureAtlas.h
@@ -17,40 +17,40 @@
 
 /** A class that implements a basic Texture Atlas */
 @interface TextureAtlas : NSObject {
-	int			totalQuads;
+	NSUInteger			totalQuads;
 	ccQuad2		*texCoordinates;
 	ccQuad3		*vertices;
 	GLushort	*indices;
 	Texture2D	*texture;
 }
 
-@property (readonly) int totalQuads;
+@property (readonly) NSUInteger totalQuads;
 @property (nonatomic,retain) Texture2D *texture;
 
 /** creates a TextureAtlas with an iname filename and with a capacity for n Quads
  * n is the number of Quads that will be rendered at once from this Atlas
  * n is the maximun number of Quads it will be able to render, but not the minimun
  */
-+(id) textureAtlasWithFile:(NSString*)file capacity: (int) n;
++(id) textureAtlasWithFile:(NSString*)file capacity: (NSUInteger) n;
 
 /** initializes a TextureAtlas with an iname filename and with a capacity for n Quads
  * n is the number of Quads that will be rendered at once from this Atlas
  * n is the maximun number of Quads it will be able to render, but not the minimun
  */
--(id) initWithFile: (NSString*) file capacity:(int)n;
+-(id) initWithFile: (NSString*) file capacity:(NSUInteger)n;
 
 /** updates a certain texture coordinate & vertex with new Quads.
  * n must be between 0 and the atlas capacity - 1
  * The default value of all of the Quads is 0,0,0,0,0,0,0,0, so this selector
  * must be called to initializes every Quad
  */
--(void) updateQuadWithTexture: (ccQuad2*) quadT vertexQuad:(ccQuad3*) quadV atIndex:(int) n;
+-(void) updateQuadWithTexture: (ccQuad2*) quadT vertexQuad:(ccQuad3*) quadV atIndex:(NSUInteger) n;
 
 
 /** draws n quads
  * n can't be greater than the capacity of the Atlas
  */
--(void) drawNumberOfQuads: (int) n;
+-(void) drawNumberOfQuads: (NSUInteger) n;
 
 /** draws all the Atlas's Quads
  */
@@ -60,6 +60,6 @@
 /** resize the capacity of the Texture Atlas.
  * The new capacity can be lower or higher
  */
--(void) resizeCapacity: (int) n;
+-(void) resizeCapacity: (NSUInteger) n;
 
 @end
